+++ examples/multi_array/main.c
@@ -23,7 +23,7 @@ void toggle (uint8_t pin) {
 
 int main(void)
 {
-    for (int i=0;i<NTASKS;i++) {
+    for (uint8_t i=0;i<NTASKS;i++) {
         pinMode(i, OUTPUT);
     }
